+++ lib/Parser/MimeDir.php
@@ -7,6 +7,7 @@ use Sabre\VObject\EofException;
 use Sabre\VObject\Component;
 use Sabre\VObject\Component\VCalendar;
 use Sabre\VObject\Component\VCard;
+use Sabre\VObject\Document;
 
 /**
  * MimeDir parser.
@@ -459,7 +460,7 @@ class MimeDir extends Parser {
             $propObj->setQuotedPrintableValue($this->extractQuotedPrintableValue());
         } else {
             $charset = $this->charset;
-            if (isset($propObj['CHARSET'])) {
+            if ($this->root->getDocumentType()===Document::VCARD21 && isset($propObj['CHARSET'])) {
                 // vCard 2.1 allows the character set to be specified per property.
                 $charset = (string)$propObj['CHARSET'];
             }